requirements.txt
python-dotenv
openai
tabulate
